Move PointsTo function to util
