--- pkg/controller/cluster/cluster_controller_test.go
@@ -1,77 +1,0 @@
-/*
-Copyright 2018 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package cluster
-
-import (
-	"testing"
-
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
-)
-
-func TestPointsTo(t *testing.T) {
-	targetID := "fri3ndsh1p"
-
-	meta := metav1.ObjectMeta{
-		UID: types.UID(targetID),
-	}
-
-	tests := []struct {
-		name     string
-		refIDs   []string
-		expected bool
-	}{
-		{
-			name: "empty owner list",
-		},
-		{
-			name:   "single wrong owner ref",
-			refIDs: []string{"m4g1c"},
-		},
-		{
-			name:     "single right owner ref",
-			refIDs:   []string{targetID},
-			expected: true,
-		},
-		{
-			name:   "multiple wrong refs",
-			refIDs: []string{"m4g1c", "h4rm0ny"},
-		},
-		{
-			name:     "multiple refs one right",
-			refIDs:   []string{"m4g1c", targetID},
-			expected: true,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			pointer := &metav1.ObjectMeta{}
-
-			for _, ref := range test.refIDs {
-				pointer.OwnerReferences = append(pointer.OwnerReferences, metav1.OwnerReference{
-					UID: types.UID(ref),
-				})
-			}
-
-			result := pointsTo(pointer.OwnerReferences, &meta)
-			if result != test.expected {
-				t.Errorf("expected %v, got %v", test.expected, result)
-			}
-		})
-	}
-}
