+++ pkg/util/util_test.go
@@ -20,6 +20,9 @@ import (
 	"io/ioutil"
 	"os"
 	"testing"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 )
 
 const validCluster = `
@@ -333,4 +336,57 @@ func createTempFile(contents string) (string, error) {
 	defer f.Close()
 	f.WriteString(contents)
 	return f.Name(), nil
+}
+
+func TestPointsTo(t *testing.T) {
+	targetID := "fri3ndsh1p"
+
+	meta := metav1.ObjectMeta{
+		UID: types.UID(targetID),
+	}
+
+	tests := []struct {
+		name     string
+		refIDs   []string
+		expected bool
+	}{
+		{
+			name: "empty owner list",
+		},
+		{
+			name:   "single wrong owner ref",
+			refIDs: []string{"m4g1c"},
+		},
+		{
+			name:     "single right owner ref",
+			refIDs:   []string{targetID},
+			expected: true,
+		},
+		{
+			name:   "multiple wrong refs",
+			refIDs: []string{"m4g1c", "h4rm0ny"},
+		},
+		{
+			name:     "multiple refs one right",
+			refIDs:   []string{"m4g1c", targetID},
+			expected: true,
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			pointer := &metav1.ObjectMeta{}
+
+			for _, ref := range test.refIDs {
+				pointer.OwnerReferences = append(pointer.OwnerReferences, metav1.OwnerReference{
+					UID: types.UID(ref),
+				})
+			}
+
+			result := PointsTo(pointer.OwnerReferences, &meta)
+			if result != test.expected {
+				t.Errorf("expected %v, got %v", test.expected, result)
+			}
+		})
+	}
 }
